% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gos.R
\name{bestkappa}
\alias{bestkappa}
\title{Function for the best kappa parameter}
\usage{
bestkappa(formula,data = NULL,kappa=seq(0.05,1,0.05),
          nrepeat = 10,nsplit = 0.5,cores = 1)
}
\arguments{
\item{formula}{A formula of GOS model}

\item{data}{A data.frame or tible of observation data}

\item{kappa}{(optional)A numeric vector of the optional percentages of observation locations
with high similarity to a prediction location.
kappa = 1 - tau, where tau is the probability parameter
in quantile operator. kappa = 0.25 means
that 25\% of observations with high similarity to a prediction
location are used for modelling.}

\item{nrepeat}{(optional)A numeric value of the number of cross-validation training times.
The default value is 10.}

\item{nsplit}{(optional)The sample training set segmentation ratio,which in \verb{(0,1)},
default is \code{0.5}.}

\item{cores}{positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster
object.}
}
\value{
A list of the result of the best kappa and the computation process curve.
}
\description{
Computationally optimized function for determining the best kappa parameter for the optimal similarity
}
\examples{
\dontrun{
library(dplyr)
library(ggplot2)
library(ggrepel)
data(zn)
data(grid)
system.time({
  b1 = bestkappa(Zn ~ Slope + Water + NDVI  + SOC + pH + Road + Mine,
                 data = zn,kappa = c(0.01, 0.05, 0.1, 0.2, 0.5, 1),
                 nrepeat = 2,cores = 1)
})
b1$bestkappa
b1$plot
}
}
\references{
Song, Y. (2022). Geographically Optimal Similarity. Mathematical Geosciences. doi: 10.1007/s11004-022-10036-8.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
