% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{gd_bestunidisc}
\alias{gd_bestunidisc}
\title{Best univariate discretize parameters based on geodetector q-statistic}
\usage{
gd_bestunidisc(
  y,
  x,
  k,
  method = c("fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust",
    "fisher", "jenks", "dpih", "headtails", "maximum", "box"),
  cores = 1,
  return_disc = TRUE,
  ...
)
}
\arguments{
\item{y}{Variable Y, continuous numeric vector.}

\item{x}{Covariate X, continuous numeric vector.}

\item{k}{A vector of number of classes for discretization.}

\item{method}{(optional) A vector of methods for discretization,default is all can
used in \code{spEcula}.}

\item{cores}{positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster
object.}

\item{return_disc}{(optional) Whether or not return discretized result used the optimal parameter.
Default is \code{TRUE}.}

\item{...}{(optional) Other arguments passed to \code{st_unidisc()}.}
}
\value{
A list with the optimal parameter in the provided parameter combination with \code{k}
and \code{method}.
}
\description{
Function for determining the best univariate discretize parameters based on geodetector q-statistic.
}
\examples{
\dontrun{

}
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
