% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosom.R
\name{geosom_bestsupperclass}
\alias{geosom_bestsupperclass}
\title{Screeplot of GeoSOM superclasses}
\usage{
geosom_bestsupperclass(
  gsom,
  nclass = 2,
  method = "pam",
  hmethod = "complete",
  bindcoord = FALSE
)
}
\arguments{
\item{gsom}{\code{kohonen} object, a SOM created by the \code{spEcula::geosom}
function.}

\item{nclass}{number of superclasses to be visualized in the screeplot.
Default is 2.}

\item{method}{Method used for clustering. Hierarchical clustering
("hclust") and Partitioning around medoids ("pam") can be used.
Default is Partitioning around medoids.}

\item{hmethod}{For hierarchicical clustering, the clustering method, by
default "complete". See the \code{stats::hclust} documentation for more
details.}

\item{bindcoord}{Does the cluster of superclass in GeoSOM consider spatial coordination.
Defaul is \code{FALSE}.}
}
\value{
No return value, called for side effects.
}
\description{
The screeplot, helps deciding the optimal number of superclasses. Available
for both PAM(\code{pam}) and hierarchical clustering(\code{hclust}).
}
\examples{
\dontrun{
data(pmc)
set.seed(2004)
g = geosom(data = pmc, coords = c("centroidx","centroidy"),
wt = 3,grid = geosomgrid(6,10),normalize = TRUE)
geosom_bestsupperclass(g)
}
}
