% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosom.R
\name{geosom_clusterlabel}
\alias{geosom_clusterlabel}
\title{Cluster label of GeoSOM}
\usage{
geosom_clusterlabel(gsom, superclass)
}
\arguments{
\item{gsom}{A \code{kohonen} object,get from \code{geosom()}.}

\item{superclass}{A numeric vector get form \code{geosom_superclass()}}
}
\value{
A numeric vector representing the cluster label.
}
\description{
Get the cluster label(result) of GeoSOM model.
}
\examples{
data(pmc)
g = geosom(data = pmc, coords = c("centroidx","centroidy"),
wt = 3,grid = geosomgrid(6,10),normalize = TRUE)
g_superclass = geosom_superclass(g,12)
g_label = geosom_clusterlabel(g,g_superclass)
g_label
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
