% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosom.R
\name{geosom_plot}
\alias{geosom_plot}
\title{Interactive GeoSOM plots}
\usage{
geosom_plot(gsom, type, superclass, ...)
}
\arguments{
\item{gsom}{A \code{kohonen} object,get from \code{geosom()}.}

\item{type}{Character, the plot type. The default "Hitmap" is a population map.
"Cloud" plots the observations as a scatterplot within each cell (see Details).
"UMatrix" plots the average distance of each cell to its neighbors, on a color scale.
"Circular" (barplot), "Barplot", "Boxplot", "Radar" and "Line" are for numeric variables.
"Color" (heat map) is for a single numeric variable. "Pie" (pie chart) and "CatBarplot"
are for a single categorical (factor) variable.}

\item{superclass}{A numeric vector get form \code{geosom_superclass()}.}

\item{...}{Other arguments passed to \code{aweSOM::aweSOMplot()}.}
}
\value{
An object of class htmlwidget.
}
\description{
Plot interactive visualizations of GeoSOM, as an html page. The plot can represent general map
informations, or selected categorical or numeric variables (not necessarily the ones used during training).
Hover over the map to focus on the selected cell or variable, and display further information.This function
is a wrapper of \code{aweSOM::aweSOMplot()}.
}
\examples{
\dontrun{
data(pmc)
set.seed(2004)
g = geosom(data = pmc, coords = c("centroidx","centroidy"),
wt = 3,grid = geosomgrid(6,10),normalize = TRUE)
g_superclass = geosom_superclass(g,12)
geosom_plot(g,type = "Circular",superclass = g_superclass)
}
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
