% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosom.R
\name{geosom_quality}
\alias{geosom_quality}
\title{GeoSOM quality measures}
\usage{
geosom_quality(gsom)
}
\arguments{
\item{gsom}{A \code{kohonen} object,get from \code{geosom()}.}
}
\value{
A list containing quality measures : quantization error, share of explained variance,
topographic error and Kaski-Lagus error.
}
\description{
Computes several quality measures on a trained GeoSOM.This function is a wrapper
of \code{aweSOM::somQuality()}.
}
\examples{
data(pmc)
set.seed(2004)
g = geosom(data = pmc, coords = c("centroidx","centroidy"),
wt = 3,grid = geosomgrid(6,10),normalize = TRUE)
geosom_quality(g)
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
