% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosom.R
\name{geosom_superclass}
\alias{geosom_superclass}
\title{Superclasses of GeoSOM}
\usage{
geosom_superclass(
  gsom,
  k,
  method = "pam",
  hmethod = "complete",
  bindcoord = FALSE
)
}
\arguments{
\item{gsom}{A \code{kohonen} object,get from \code{geosom()}.}

\item{k}{The number of superclasses.}

\item{method}{(optional)The clustering algorithms used on the map’s prototypes,two methods are
implemented in \code{spEcula}, PAM(k-medians) and hierarchical clustering.When method is \code{pam},PAM
(k-medians) is used,otherwise hierarchical clustering.Default is \code{pam}.}

\item{hmethod}{For hierarchicical clustering, the clustering method, by
default "complete". See the \code{stats::hclust} documentation for more
details.}

\item{bindcoord}{Does the cluster of superclass in GeoSOM consider spatial coordination.
Defaul is \code{FALSE}.}
}
\value{
A numeric vector representing the superclass.
}
\description{
It is common to further cluster the GeoSOM map into superclasses, groups of cells with similar
profiles. This is done using classic clustering algorithms on the map’s prototypes.Remeber to
use it when the non-spatial variable is more than 1.
}
\examples{
data(pmc)
set.seed(2004)
g = geosom(data = pmc, coords = c("centroidx","centroidy"),
wt = 3,grid = geosomgrid(6,10),normalize = TRUE)
g_superclass = geosom_superclass(g,12)
g_superclass
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
