% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosom.R
\name{geosomgrid}
\alias{geosomgrid}
\title{GeoSOM-grid related functions}
\usage{
geosomgrid(
  xdim,
  ydim,
  topo = "hexagonal",
  neighbourhood.fct = "gaussian",
  toroidal = FALSE
)
}
\arguments{
\item{xdim}{X dimensions of the grid}

\item{ydim}{Y dimensions of the grid.}

\item{topo}{(optional)Choose between a \code{hexagonal} or \code{rectangular} topology,default is \code{hexagonal}.}

\item{neighbourhood.fct}{(optional)Choose between \code{bubble} and \code{gaussian} neighbourhoods when training a
GeoSOM."bubble" or "gaussian"(default).}

\item{toroidal}{(optional)Logical, whether the grid is \code{toroidal} or not.Default is \code{FALSE}.}
}
\value{
An object of class "somgrid", with elements pts, and the input arguments to the function.
}
\description{
Set up a grid of units, of a specified size and topology.Wrapper of \code{kohonen::somgrid}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
