% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmc}
\alias{pmc}
\title{Portugal municipal census data 2011}
\format{
\code{pmc}: A tibble of 2011 Portugal Census data with 278 rows and 14 variables.
}
\usage{
pmc
}
\description{
The 2011 Portugal Census data contains the centroid coordinates and
12 selected socio-economic variables for the 278 municipalities of Portugal.The
12 selected variables are \code{denspop11}(Population density) \code{discapd}(Average distance to the closest district capital)
\code{gmm2011}(Average monthly earnings) \code{podc}(Daily purchasing power per capita) \code{tdesmp11}(Unemployment rate)
\code{pbsdt11}(Population receiving the unemployment benefit) \code{ne_sup}(Population with a university degree completed)
\code{abanesc}(Percentage of school dropout) \code{pdmcdt11}(Expenses of the municipality on cultural and sports activities)
\code{pdiv}(Percentage of divorces) \code{distpolall}(Average distance to the closest police station) \code{tcrmiltot}(Criminality rate)
}
\keyword{2011}
\keyword{Census}
\keyword{Portugal}
\keyword{dataset}
