% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{sandwich}
\alias{sandwich}
\title{sandwich mapping model}
\usage{
sandwich(
  sampling,
  stratification,
  reporting,
  sampling_attr,
  ssh_zone,
  reporting_id,
  weight_type = "area"
)
}
\arguments{
\item{sampling}{Sampling layer, spatial point vector object which is \code{sf} or can be converted to \code{sf} object.}

\item{stratification}{Stratification layer, spatial polygon vector object which is \code{sf} or can be converted to \code{sf} object.}

\item{reporting}{Reporting layer, spatial polygon vector object which is \code{sf} or can be converted to \code{sf} object.}

\item{sampling_attr}{The \code{attribute} column for the sampling point in sampling layer.}

\item{ssh_zone}{The \code{zone} column for the stratification layer.}

\item{reporting_id}{The \code{id} column for the reporting layer.}

\item{weight_type}{(optional) Geographic area based on weight(\code{area}) or indicate human population size(\code{population}) ,
Default is \code{area}.}
}
\value{
A \code{sf} object with estimated mean \code{sandwichest_mean} and standard error \code{sandwichest_standarderror}.
}
\description{
Spatial prediction based on spatial stratified heterogeneity using sandwich mapping model.
}
\examples{
library(sf)
simpath = system.file("extdata", "sim.gpkg", package="spEcula")
sampling = read_sf(simpath,layer = 'sim_sampling')
ssh = read_sf(simpath,layer = 'sim_ssh')
reporting = read_sf(simpath,layer = 'sim_reporting')
sandwich(sampling = sampling,stratification = ssh,reporting = reporting,
        sampling_attr = 'Value',ssh_zone = 'X',reporting_id = 'Y',
        weight_type = 'population')
sandwich(sampling = sampling,stratification = ssh,reporting = reporting,
        sampling_attr = 'Value',ssh_zone = 'X',reporting_id = 'Y',
        weight_type = 'area')
}
\references{
Lin, Y., Xu, C., & Wang, J. (2023). sandwichr: Spatial prediction in R based on spatial stratified heterogeneity.
Transactions in GIS: TG, 27(5), 1579–1598. https://doi.org/10.1111/tgis.13088
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
