% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gos.R
\name{gos}
\alias{gos}
\title{Geographically optimal similarity}
\usage{
gos(formula, data = NULL, newdata = NULL, kappa = 0.25, cores = 1)
}
\arguments{
\item{formula}{A formula of GOS model.}

\item{data}{A data.frame or tibble of observation data.}

\item{newdata}{A data.frame or tibble of prediction variables data.}

\item{kappa}{A numeric value of the percentage of observation locations
with high similarity to a prediction location.
kappa = 1 - tau, where tau is the probability parameter
in quantile operator. The default kappa is 0.25, meaning
that 25\% of observations with high similarity to a prediction
location are used for modelling.}

\item{cores}{positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster object}
}
\value{
A tibble made up of predictions and uncertainties.
}
\description{
Computationally optimized function for geographically optimal similarity (GOS) model
}
\examples{
\dontrun{
g = gos(Zn ~ Slope + Water + NDVI  + SOC + pH + Road + Mine,
data = zn, newdata = grid, kappa = 0.08,cores = 6)
g
}
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
