% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid}
\alias{grid}
\title{Spatial grid data of explanatory variables}
\format{
\code{grid}: A tibble of grided trace element explanatory variables
with 13132 rows and 12 variables, where the first column is GridID.
}
\usage{
grid
}
\description{
Spatial grid data of explanatory variables,modified from \code{geosimilarity}
package.
}
\keyword{dataset}
\keyword{elements}
\keyword{trace}
